+++ src/Util/Filesystem/UserPaths.hpp
@@ -3,7 +3,7 @@
 #include <array>
 #include <string_view>
 #include <filesystem>
-#include "Util/Filesystem/API.hpp"
+#include "Util/Filesystem/FilesystemAPI.hpp"
 
 namespace VSCL::FS {
